Correction to MEAN and MEDIAN score generation.
I noticed there was a small issue with the computation of mean and median error. With the published version, mean and median will report LOGGED mean and median. I changed a couple of lines to report mean and median BEFORE logging the figures. If you need the old figures, then change back. Thanks